backend/requirements.txt
Django==5.0.6
djangorestframework==3.15.2
django-cors-headers==4.4.0
django-filter==24.2
dj-database-url==2.2.0
psycopg2-binary==2.9.9
gunicorn==22.0.0
whitenoise==6.7.0
ruff==0.5.5
